Refactor auth-related functions.
