+++ app/lib/frontend/service_utils.dart
@@ -11,25 +11,15 @@ import 'package:appengine/appengine.dart';
 import 'package:gcloud/db.dart';
 import 'package:gcloud/service_scope.dart';
 import 'package:http/http.dart' as http;
-import 'package:shelf/shelf.dart' as shelf;
 
 import '../shared/configuration.dart';
 import '../shared/package_memcache.dart';
 
 import 'backend.dart';
-import 'oauth2_service.dart';
 import 'search_service.dart';
 import 'upload_signer_service.dart';
 
 final String templatePath = Platform.script.resolve('../views').toFilePath();
-
-void initOAuth2Service() {
-  // The oauth2 service is used for getting an email address from an oauth2
-  // access token (which the pub client sends).
-  final client = new http.Client();
-  registerOAuth2Service(new OAuth2Service(client));
-  registerScopeExitCallback(() async => client.close());
-}
 
 void initSearchService() {
   registerSearchService(new SearchService());
@@ -40,23 +30,6 @@ void initBackend(
     {UIPackageCache cache, FinishedUploadCallback finishCallback}) {
   registerBackend(new Backend(dbService, tarballStorage,
       cache: cache, finishCallback: finishCallback));
-}
-
-/// Looks at [request] and if the 'Authorization' header was set tries to get
-/// the user email address and registers it.
-Future registerLoggedInUserIfPossible(shelf.Request request) async {
-  final authorization = request.headers['authorization'];
-  if (authorization != null) {
-    final parts = authorization.split(' ');
-    if (parts.length == 2 && parts.first.trim().toLowerCase() == 'bearer') {
-      final accessToken = parts.last.trim();
-
-      final email = await oauth2Service.lookup(accessToken);
-      if (email != null) {
-        registerLoggedInUser(email);
-      }
-    }
-  }
 }
 
 Future<String> obtainServiceAccountEmail() async {
